requirements.txt
python-docx
pypandoc
pymupdf
striprtf
tqdm
catdoc